backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.0.1
redis==5.0.1 